Reusing the GlassUpAgent class for bug fixing
the GlassUpAgent class from the library calls Intents
in a wise which is not supported in Android version 5
and above.
Extending the class is not possible as all attributes
are private and couldn't be used within the derived class
so I just implemented a complete new class with the same
code as the GlassUpAgent class and fixed the bug within
my own class.

The Voice Recorder cannot be initialized because of
the error:
AudioRecord: AudioFlinger could not create record track, status: -1
this will be adressed next