GoogleSpeechRecognition Voice Recorder and Stream Client
Got the Voice Recorder to work

recording through instance of AudioRecord in
VoiceRecorder.java now and storing the raw
bytes in a queue within the GoogleRecorder.java

the GoogleStreamClient.java creates a managed channel
to speech api. A response is received, the authentication
needs to be fixed. After this the records from the queue
can be send.

Also fixed the package names to fit better to the project

renamed Recorder interface to IRecorder

created an interface for the AR Printers for new printers
to be added easyly

also created a printerFactory returning one of the printers

in future a combo box should exist, from where the user has
the ability to decide, which printer / speech recognition
should be used

missing: code documentation. Will be fixed in next commit